Moved Checkbox, SearchField to it's own components
