+++ src/components/Checkbox/index.js
@@ -1,16 +1,18 @@
-const Checkbox = ({ label, name, onChange, checked, value }) => {
+import "./checkbox.css";
+
+const Checkbox = ({ label, id, name, onChange, checked, value }) => {
   return (
     <div className="checkbox-wrapper flex">
       <input
         className="custom-checkbox"
         type="checkbox"
         name={name}
-        id={name}
+        id={id}
         onChange={onChange}
         checked={checked}
         value={value}
       />
-      <label htmlFor={name}>{label}</label>
+      <label htmlFor={id}>{label}</label>
     </div>
   );
 };
