+++ src/components/SearchField/search.css
@@ -1,0 +1,20 @@
+/* Search */
+.search-wrapper {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+
+  color: var(--border);
+  border: 1px solid var(--border);
+  border-radius: 4px;
+  padding-left: 4px;
+}
+
+.search-wrapper input {
+  width: 100%;
+  border: none;
+}
+/* Magnifying glass icon */
+.search-wrapper i {
+  font-size: 14px;
+}
